install libraries for each os

+++ ruby_setup.sh
@@ -17,7 +17,12 @@ exec $SHELL -l
 git clone https://github.com/rbenv/ruby-build.git ~/.rbenv/plugins/ruby-build
 
 # install tools for ruby
-sudo yum install -y openssl-devel readline-devel zlib-devel
+if [ -e /etc/debian-version ]; then
+  sudo apt-get update
+  sudo apt-get install -y libssl-dev libreadline-dev zlib1g-dev
+elif [ -e /etc/redhat-release ]; then
+  sudo yum install -y openssl-devel readline-devel zlib-devel
+fi
 
 # Then, run under commands.
 # $ rbenv install {version}
